Fucked shit up big time
What the title says.
So I wanted to truly represent the logic of the shell, as in I wanted to
make sure that the waitpid() calls were made to the proper waitpids, and
not with (0) as a parameter. As a result I now seg fault, and I still do
not know how to debug programs with forks....

So I stashed my changes in this branch, but this is not over!

Here's how to fix:
- Add an array in the main container that contains all the waitpids.
- Dynamically adjust the number of pids depending on whether the first
and last commands fail or succeed.

Here's how to test :
- [ ] /bin/cat /dev/urandom should not hang, but also should not even start if
the in_fd is wrong.
- Convertly, the last command should never get executed if the
corresponding out_fd == -1